smart-waste-management/requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
Pillow==9.5.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-tables2==2.5.3
gunicorn==20.1.0
whitenoise==6.4.0
python-dotenv==1.0.0
django-storages==1.13.2
boto3==1.26.137
django-filter==23.2
django-widget-tweaks==1.4.12 